requirements.txt
flask==2.3.3
flask-cors==5.0.0
feedparser==6.0.10
requests==2.32.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
httpx==0.28.1
jinja2==3.1.2
beautifulsoup4==4.12.2
schedule==1.2.0
supabase==2.28.3
websockets==13.0
